Completes the four-branch integration: master + cast_relabel (incl. best_of_n)
+ surya/rl-token + routing-commands.

Resolved 45 hunks across 10 files.

* impls/jax_agents/sac_residual.py -> SPLIT. The two branches added this file
  independently with the same class names but incompatible roles:
    - jax_agents/sac_residual.py (surya/rl-token) keeps the name. It is a
      standalone SAC -- own critic, StateHead, select_action_otf -- registered
      as agents["sac_residual"] and selected via agent_name.
    - jax_agents/dsrl_residual.py (NEW, routing-commands) is DSRL's residual
      sub-agent: no critic of its own, adds the DAgger paths and obs
      normalization, selected via agent_name="dsrl" +
      online_training_mode="sac_residual"/"dagger_residual".
  dsrl.py, main_carla.py and the DELTA_XY docstring reference now point at the
  right module; a header on dsrl_residual.py explains the split.

* run_carla.sh -> rebuilt from routing-commands' (the larger base) plus dev's
  six flags. Verified a strict superset of both parents: 52 flags, none
  dropped. Default base agent config is steervla_dsrl_config.py (both residual
  stacks are opt-in via --agent-config), and the crash-supervisor loop
  (--max-retries + --resume) is carried over from surya/rl-token.
  Also made two hardcoded paths non-fatal: CARLA_ROOT now warns instead of
  silently substituting another box's path, and the CARLA 0.9.15 override is
  opt-in via CARLA_0915_ROOT rather than an unconditional hard exit.

* main_carla.py -> 21 hunks. main() keeps surya's dispatcher (sac_residual ->
  _run_residual_entry); _run_dsrl_entry is routing-commands' session-based body
  (build_carla_session + run_online_carla) grafted onto master's richer
  exp_name construction. The video path is routing-commands' throughout
  (_annotate_candidates_panel / _annotate_text_panel / _save_local_video) with
  master's step_in_video + episode_video_frame_index bookkeeping, which
  _append_trajectory_step and the cast_relabel/VLM coaches depend on. step_wb
  uses _reward_breakdown_log plus routing-commands' bon/* and violation keys.
  Update gating is the union: any_updates_on and not expert_debug and not
  eval_only.

* Several bulk unions interleaved two functions mid-body and had to be
  de-interleaved by hand (_plot_bon_candidates vs the annotation helpers,
  _save_local_video vs _append_trajectory_step). Pieces lost in that process
  and restored: _as_video_frame, episode_trajectory / episode_video_fps /
  episode_video_frame_index, and the openpi `pi0 as _openpi_pi0` import that
  routing-commands' QGF helpers need.

* carla_utils.py -> kept the portable _resolve_nvidia_vk_icd() over
  routing-commands' hardcoded /etc/vulkan path, and routing-commands'
  pass-through of all PID decoder kwargs.
* critic_feedback.py -> both guards, with the critic_pretrained_weights ->
  "none" override checked first (it is a hard override, not a preference).
* steervla.py -> QGF-guided denoise and CSP t_context coexist: QGF runs its own
  denoising loop, otherwise sampling goes through the cached path with
  t_context.
* steervla_dsrl_config.py -> single active checkpoint, matching actor_config
  (..._no_ego_history).

ruff F821/F811 across the touched files: 13, below both parents (dev 24,
rc-integrate 3+); every remaining one is pre-existing (surya's commented-out
flow_sample body and its duplicate sample_candidates, routing-commands'
_openpi_gemma, master's duplicate warn_if_carla_root_mismatched).

+++ impls/coaches/simlingo/autopilot.py
@@ -2006,7 +2006,7 @@ class AutoPilot(autonomous_agent_local.AutonomousAgent):
     angle_radians = -math.atan2(-aim_y, aim_x)
 
     # Convert the angle from radians to degrees
-    angle_degrees = np.float_(math.degrees(angle_radians))
+    angle_degrees = float(math.degrees(angle_radians))
 
     return angle_degrees
